Flyttet skjemavalidering til egen funksjon som kan brukes på tvers
